[REFACTOR/#8] presentation 액티비티별 분리
