+++ app/src/main/res/layout/fragment_home.xml
@@ -5,7 +5,7 @@
 
         <variable
             name="vm"
-            type="org.sopt.sample.presentation.home.HomeViewModel" />
+            type="org.sopt.sample.presentation.main.home.HomeViewModel" />
     </data>
 
     <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
@@ -14,7 +14,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:paddingHorizontal="20dp"
-        tools:context=".presentation.home.HomeFragment">
+        tools:context=".presentation.main.home.HomeFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_follower"
