+++ app/src/main/res/layout/fragment_setting.xml
@@ -11,7 +11,7 @@
         android:id="@+id/frame"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.setting.SettingFragment">
+        tools:context=".presentation.main.setting.SettingFragment">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/layout"
