+++ app/src/main/res/layout/fragment_user.xml
@@ -11,7 +11,7 @@
         android:id="@+id/frame"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.user.UserFragment">
+        tools:context=".presentation.main.user.UserFragment">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/layout"
